New alloc-free incremental hash (#1)
Merged over [twmb/murmur3](https://github.com/twmb/murmur3) as v2, since the APIs are different, compared to the older library and it has very good non-incremental hashing functions.

Also, separately created a module of incremental hashes as `v2/stackmurmur3` with slightly different API to match `twmb/murmur3`. However, due to lack of alignment-unsafe code, it is expectedly slower for large blocks of data. The compiler is performing limited optimizations, so the only alternative is to drop down to assembly level - which `twmb/murmur3` does for non-streaming 128-bit murmur3 variant on amd64.